joined logout tests relates #15
